Remove the legacy Python config implementation.
