+++ src/python/pants/help/help_info_extracter_test.py
@@ -2,11 +2,11 @@
 # Licensed under the Apache License, Version 2.0 (see LICENSE).
 
 from enum import Enum
-from types import SimpleNamespace
 from typing import Any, Iterable, List, Optional, Tuple, Union
 
 from pants.base.build_environment import get_buildroot
 from pants.build_graph.build_configuration import BuildConfiguration
+from pants.engine.fs import FileContent
 from pants.engine.goal import GoalSubsystem
 from pants.engine.internals.parser import BuildFileSymbolInfo, BuildFileSymbolsInfo
 from pants.engine.rules import collect_rules, rule
@@ -257,7 +257,7 @@ def test_get_all_help_info(tmp_path) -> None:
         core_fields = (QuxField, QuuxField)
 
     config_path = "pants.test.toml"
-    config_source = SimpleNamespace(path=config_path, content=b"[GLOBAL]\nopt1 = '+[99]'")
+    config_source = FileContent(path=config_path, content=b"[GLOBAL]\nopt1 = '+[99]'")
     options = Options.create(
         args=["./pants", "--backend-packages=['internal_plugins.releases']"],
         env={"PANTS_OPT1": "88"},
